requirements.txt
streamlit
opencv-python
pytesseract
pandas
pillow
numpy
